@@ -29,10 +29,10 @@
     "@types/enzyme-adapter-react-16": "^1.0.3",
     "@types/history": "4.7.2",
     "@types/node": "10.12.10",
-    "@types/react": "16.8.23",
+    "@types/react": "16.9.2",
     "@types/react-collapse": "^4.0.2",
     "@types/react-document-title": "^2.0.3",
-    "@types/react-dom": "^16.8.5",
+    "@types/react-dom": "16.9.0",
     "@types/react-redux": "7.1.1",
     "@types/react-router-dom": "4.3.5",
     "@types/styled-components": "4.1.8",
@@ -47,7 +47,7 @@
     "case-sensitive-paths-webpack-plugin": "2.1.2",
     "chalk": "2.4.1",
     "classnames": "^2.2.6",
-    "connected-react-router": "4.5.0",
+    "connected-react-router": "6.5.2",
     "css-loader": "1.0.0",
     "dotenv": "6.0.0",
     "dotenv-expand": "4.2.0",
@@ -102,12 +102,12 @@
     "promise": "8.0.1",
     "promise-polyfill": "6.0.2",
     "prop-types": "^15.7.2",
-    "react": "16.8.6",
+    "react": "16.9.0",
     "react-app-polyfill": "^0.1.3",
     "react-collapse": "^4.0.3",
     "react-dev-utils": "^6.1.1",
     "react-document-title": "2.0.3",
-    "react-dom": "16.8.6",
+    "react-dom": "16.9.0",
     "react-modal": "2.2.4",
     "react-motion": "^0.5.2",
     "react-redux": "7.1.0",
@@ -115,7 +115,7 @@
     "react-scroll": "1.7.10",
     "redux": "4.0.4",
     "redux-devtools-extension": "2.13.5",
-    "redux-saga": "0.16.2",
+    "redux-saga": "1.0.5",
     "resolve": "1.8.1",
     "sass-loader": "7.1.0",
     "scroll-behavior": "0.3.2",
@@ -173,7 +173,7 @@
     "typescript": "3.1.3",
     "webpack": "4.19.1",
     "webpack-cli": "3.1.0",
-    "webpack-dev-server": "3.1.9"
+    "webpack-dev-server": "^3.1.11"
   },
   "eslintConfig": {
     "extends": "react-app"
