@@ -1,10 +1,10 @@
 import { Store, createStore, applyMiddleware } from 'redux';
 import createSagaMiddleware from 'redux-saga';
-import { connectRouter, routerMiddleware } from 'connected-react-router';
+import { routerMiddleware } from 'connected-react-router';
 import { composeWithDevTools } from 'redux-devtools-extension';
-import { History } from 'history';
 
 import { ApplicationState, rootReducer, rootSaga } from './index';
+import { History } from 'history';
 
 export default function configureStore(
   history: History,
@@ -14,7 +14,7 @@ export default function configureStore(
   const sagaMiddleware = createSagaMiddleware();
 
   const store = createStore(
-    connectRouter(history)(rootReducer),
+    rootReducer(),
     initialState,
     composeEnhancers(applyMiddleware(routerMiddleware(history), sagaMiddleware))
   );
